// lib/movie/movie.types.ts
export type Movie = {
  imdbId: string;
  title: string;
  year: string;
  rated: string | null;
  released: string | null;
  runtime: string | null;
  genre: string | null;
  director: string | null;
  actors: string | null;
  plot: string | null;
  posterUrl: string | null;
  imdbRating: string | null;
};